Fixed unsigned/signed comparison warning.
Also fixed LDFLAG ordering issue.

+++ src/parser.cpp
@@ -92,7 +92,7 @@ PixelSortQuery_t * process_tokens(const char* query_string) {
 }
 
 void destroy_query(PixelSortQuery_t * query) {
-    for(int i = 0; i < query->subquery_count; ++i) {
+    for(size_t i = 0; i < query->subquery_count; ++i) {
 	free(query->subqueries[i]);
     }
     free(query);
